MTD Classification model/requirements.txt
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==43.0.0
cuda-python==12.5.0
cycler==0.12.1
datumaro==1.6.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dnspython==2.6.1
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
geti-sdk==2.2.0
gevent==24.2.1
geventhttpclient==2.0.2
greenlet==3.0.3
grpcio==1.65.1
h5py==3.11.0
huggingface-hub==0.24.3
idna==3.7
imageio-ffmpeg==0.5.1
imagesize==1.4.1
ipython==8.18.1
jedi==0.19.1
joblib==1.4.2
json-stream==2.3.2
json-stream-rs-tokenizer==0.4.26
jstyleson==0.0.2
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
multidict==6.0.5
networkx==2.8
nibabel==5.2.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
openvino==2024.2.0
openvino-dev==2024.2.0
openvino-telemetry==2024.1.0
openvino_model_api==0.2.1
orjson==3.10.0
otx==1.6.2
ovmsclient==2023.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathvalidate==3.2.0
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycocotools==2.0.8
pycparser==2.22
pyemd==1.0.0
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.2
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-rapidjson==1.19
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.13.1
semver==3.0.2
Shapely==1.8.0
simplejson==3.19.2
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tensorboardX==2.6.2.2
texttable==1.7.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
traitlets==5.14.3
tritonclient==2.41.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.2.13
XlsxWriter==3.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
